backend/app/requirements.txt
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==23.2.0
black==24.4.2
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mypy-extensions==1.0.0
openai==1.54.3
packaging==24.1
pathspec==0.12.1
pdoc==15.0.0
platformdirs==4.2.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
rpds-py==0.19.1
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.67.0
typing_extensions==4.12.2
Werkzeug==3.0.3
